@@ -17,13 +17,14 @@
             "HyperfTest\\": "tests"
         }
     },
+    "minimum-stability": "dev",
     "require": {
-        "php": ">=7.3",
-        "hyperf/config": "^2.2",
-        "hyperf/db-connection": "^2.2",
-        "hyperf/di": "^2.2",
-        "hyperf/framework": "2.2.*",
-        "hyperf/validation": "^2.2"
+        "php": ">=8.0",
+        "hyperf/config": "~3.0.x-dev",
+        "hyperf/db-connection": "~3.0.x-dev",
+        "hyperf/di": "~3.0.x-dev",
+        "hyperf/framework": "~3.0.x-dev",
+        "hyperf/validation": "~3.0.x-dev"
     },
     "require-dev": {
         "friendsofphp/php-cs-fixer": "^3.0",
