@@ -2,41 +2,21 @@
 
 namespace DeathSatan\Hyperf\Validate\Annotation;
 
+use Attribute;
 use Hyperf\Di\Annotation\AbstractAnnotation;
 use Hyperf\Di\Exception\NotFoundException;
 
-/**
- * @Annotation
- * @Target({"CLASS", "METHOD"})
- */
+#[Attribute(Attribute::TARGET_CLASS)]
 class Validate extends AbstractAnnotation
 {
-    /**
-     * @var string 验证器类名
-     */
-    public $validate;
 
     /**
-     * @var string 场景名
+     * Validate constructor.
+     * @param string $validate 验证器类名
+     * @param array|string $scene 验证场景
      */
-    public $scene = null;
-
-
-    /**
-     * @throws NotFoundException
-     */
-    public function __construct(...$value)
+    public function __construct(public string $validate,public array|string $scene = [])
     {
-        $formattedValue = $this->formatParams($value);
-        foreach ($formattedValue as $key => $val) {
-            if ($key === 'validate')
-            {
-                $this->checkValidate($val);
-            }
-            if (property_exists($this, $key)) {
-                $this->{$key} = $val;
-            }
-        }
     }
 
     /**
